docs(documentation): 更新缓存管理规范及新增编译错误分析报告
- 补充Java标准库javax包的例外说明，明确允许使用的javax包列表
- 引入三级缓存架构设计，详细说明L1本地缓存、L2 Redis分布式缓存和L3网关缓存
- 统一缓存管理策略，推广使用UnifiedCacheManager替代自定义缓存管理器
- 调整缓存配置规范，新增application.yml示例及缓存领域配置
- 明确缓存实施要求及三级缓存优势，剔除禁止自定义缓存管理器的旧规范
- 添加完整的编译错误根因分析与修复策略报告，覆盖架构设计、Jakarta EE迁移等核心问题
- 附加修复方案分阶段计划、团队协作规范及培训计划，提升代码质量和架构合规性

+++ microservices/microservices-common-permission/src/main/java/net/lab1024/sa/common/permission/alert/PermissionAlertManager.java
@@ -1,7 +1,7 @@
 package net.lab1024.sa.common.permission.alert;
 
 import lombok.extern.slf4j.Slf4j;
-import net.lab1024.sa.common.permission.audit.PermissionAuditDTO;
+import net.lab1024.sa.common.permission.domain.dto.PermissionAuditDTO;
 import net.lab1024.sa.common.permission.audit.PermissionAuditLogger;
 import jakarta.annotation.Resource;
 import org.springframework.data.redis.core.RedisTemplate;
@@ -14,6 +14,8 @@ import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.stream.Collectors;
+import java.util.UUID;
+import java.util.HashMap;
 
 /**
  * 权限异常访问告警管理器
@@ -637,7 +639,7 @@ public class PermissionAlertManager {
             stats.setTotalAlerts(stats.getUserFailureAlerts() + stats.getIpAbnormalAlerts());
 
             // 设置当前活跃告警数
-            stats.setActiveAlerts(alertHistory.size());
+            stats.setActiveAlerts((long) alertHistory.size());
 
         } catch (Exception e) {
             log.error("[权限告警] 获取告警统计信息异常", e);
